CLOUDP-68878: split the description package
