+++ internal/cli/alerts/acknowledge.go
@@ -20,7 +20,6 @@ import (
 
 	"github.com/mongodb/mongocli/internal/cli"
 	"github.com/mongodb/mongocli/internal/config"
-	"github.com/mongodb/mongocli/internal/description"
 	"github.com/mongodb/mongocli/internal/flag"
 	"github.com/mongodb/mongocli/internal/output"
 	"github.com/mongodb/mongocli/internal/store"
@@ -74,7 +73,7 @@ func AcknowledgeBuilder() *cobra.Command {
 	opts := new(AcknowledgeOpts)
 	cmd := &cobra.Command{
 		Use:     "acknowledge <ID>",
-		Short:   description.AcknowledgeAlerts,
+		Short:   AcknowledgeAlerts,
 		Aliases: []string{"ack"},
 		Args:    cobra.ExactArgs(1),
 		PreRunE: func(cmd *cobra.Command, args []string) error {
