+++ internal/cli/alerts/description.go
@@ -11,27 +11,14 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
 package alerts
 
-import (
-	"github.com/mongodb/mongocli/internal/description"
-	"github.com/spf13/cobra"
+const (
+	Alerts              = "Manage alerts for your project."
+	AcknowledgeAlerts   = "Acknowledge an alert for your project."
+	UnacknowledgeAlerts = "Unacknowledge an alert for your project."
+	DescribeAlert       = "Describe an alert for a project."
+	ListAlerts          = "List alerts for a project."
+	Global              = "Manage Ops Manager global properties."
+	ListGlobalAlerts    = "List global alerts."
 )
-
-func ConfigBuilder() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:     "settings",
-		Aliases: []string{"settings", "config"},
-		Short:   description.Config,
-		Long:    description.ConfigLong,
-	}
-
-	cmd.AddCommand(ConfigCreateBuilder())
-	cmd.AddCommand(ConfigListBuilder())
-	cmd.AddCommand(ConfigDeleteBuilder())
-	cmd.AddCommand(ConfigsFieldsBuilder())
-	cmd.AddCommand(ConfigUpdateBuilder())
-
-	return cmd
-}
