+++ internal/cli/alerts/global_list.go
@@ -17,7 +17,6 @@ package alerts
 import (
 	"github.com/mongodb/mongocli/internal/cli"
 	"github.com/mongodb/mongocli/internal/config"
-	"github.com/mongodb/mongocli/internal/description"
 	"github.com/mongodb/mongocli/internal/flag"
 	"github.com/mongodb/mongocli/internal/output"
 	"github.com/mongodb/mongocli/internal/store"
@@ -60,7 +59,7 @@ func GlobalListBuilder() *cobra.Command {
 	opts := &GlobalListOpts{}
 	cmd := &cobra.Command{
 		Use:     "list",
-		Short:   description.ListGlobalAlerts,
+		Short:   ListGlobalAlerts,
 		Aliases: []string{"ls"},
 		Args:    cobra.NoArgs,
 		PreRunE: func(cmd *cobra.Command, args []string) error {
