+++ internal/cli/alerts/settings/description.go
@@ -1,0 +1,25 @@
+// Copyright 2020 MongoDB Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+package settings
+
+const (
+	Config             = "Manage alerts configuration for a project."
+	ConfigLong         = "The settings command provides access to your alerts configurations. You can create, edit, and delete alert configurations."
+	CreateAlertsConfig = "Create an alert configuration for a project."
+	DeleteAlertsConfig = "Delete an alert configuration from a project."
+	ConfigFields       = "Manage alert configuration fields for a project."
+	ConfigFieldsType   = "List alert configurations available field types."
+	ListConfigs        = "List alert configurations for a project."
+	UpdateAlertsConfig = "Update an alert configuration for a project."
+)
