+++ internal/cli/alerts/unacknowledge.go
@@ -17,7 +17,6 @@ package alerts
 import (
 	"github.com/mongodb/mongocli/internal/cli"
 	"github.com/mongodb/mongocli/internal/config"
-	"github.com/mongodb/mongocli/internal/description"
 	"github.com/mongodb/mongocli/internal/flag"
 	"github.com/mongodb/mongocli/internal/output"
 	"github.com/mongodb/mongocli/internal/store"
@@ -63,7 +62,7 @@ func UnacknowledgeBuilder() *cobra.Command {
 	opts := new(UnacknowledgeOpts)
 	cmd := &cobra.Command{
 		Use:     "unacknowledge <ID>",
-		Short:   description.UnacknowledgeAlerts,
+		Short:   UnacknowledgeAlerts,
 		Aliases: []string{"unack"},
 		Args:    cobra.ExactArgs(1),
 		PreRunE: func(cmd *cobra.Command, args []string) error {
