+++ internal/description/description.go
@@ -15,19 +15,6 @@ package description
 
 const (
 	Atlas                       = "Atlas operations."
-	Alerts                      = "Manage alerts for your project."
-	AcknowledgeAlerts           = "Acknowledge an alert for your project."
-	UnacknowledgeAlerts         = "Unacknowledge an alert for your project."
-	Config                      = "Manage alerts configuration for your project."
-	ConfigLong                  = "The configs command provides access to your alerts configurations. You can create, edit, and delete alert configurations."
-	CreateAlertsConfig          = "Create an alert configuration for your project."
-	DeleteAlertsConfig          = "Delete an alert configuration from your project."
-	AlertsConfigFields          = "Manage alert configuration fields for your project."
-	AlertsConfigFieldsType      = "List alert configurations available field types."
-	ListAlertsConfigs           = "List alert configurations for a project."
-	UpdateAlertsConfig          = "Update an alert configuration for a project."
-	DescribeAlert               = "Describe an alert for a project."
-	ListAlerts                  = "List alerts for a project."
 	Backup                      = "Manage backups for your project."
 	Checkpoints                 = "Manage backup checkpoints for your project."
 	ListCheckpoints             = "List continuous backup checkpoints for your project."
@@ -139,11 +126,9 @@ const (
 	DownloadLogCollectionJob    = "Download logs generated by a log collection job."
 	OpsManager                  = "Ops Manager operations."
 	Agents                      = "Manage Ops Manager agents."
-	ListGlobalAlerts            = "List global alerts."
 	Automation                  = "Manage Ops Manager automation config."
 	ShowAutomationStatus        = "Show the current status of the automation config."
 	WatchAutomationStatus       = "Watch for automation changes."
-	Global                      = "Manage Ops Manager global properties."
 	Owner                       = "Manage Ops Manager owners."
 	CreateOwner                 = "Create the first user for Ops Manager."
 	Servers                     = "Manage Ops Manager servers."
